Add custom exception handling for update and create user

+++ src/main/java/org/project/springboot_jpa/services/UserService.java
@@ -1,9 +1,13 @@
 package org.project.springboot_jpa.services;
 
+import jakarta.persistence.EntityNotFoundException;
 import org.project.springboot_jpa.entities.User;
 import org.project.springboot_jpa.repositories.UserRepository;
+import org.project.springboot_jpa.services.exceptions.DatabaseException;
 import org.project.springboot_jpa.services.exceptions.ResourceNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
+import org.springframework.dao.EmptyResultDataAccessException;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -29,13 +33,26 @@ public class UserService {
     }
 
     public void delete(Long id) {
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException(id);
+        }
+
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e) {
+            throw new DatabaseException(e.getMessage());
+        }
     }
 
     public User update(Long id, User user) {
-        User entity = repository.getReferenceById(id);
-        updateDate(entity, user);
-        return repository.save(entity);
+        try {
+            User entity = repository.getReferenceById(id);
+            updateDate(entity, user);
+            return repository.save(entity);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException(id);
+        }
+
     }
 
     private void updateDate(User entity, User user) {
